chore: Rename project to 'settable.nvim'
Naming things is tough. We don't want to overthink it, but also want the
name generic and descriptive enough to convey what's going on.

The core functionality is "applying settings" via "lua tables". You are
"setting" with "tables". Incidentally, everything you define is also
"settable", am I right?

It might be _too clever_, but it's also basic enough that I'm okay to
leave it as is. If nothing else, the lowercase is closer to neovim
plugin conventions.